requirements.txt
pillow>=10.0.0
pytesseract>=0.3.10
PyQt6>=6.5.0
google-cloud-translate>=3.12.0
googletrans==4.0.0rc1
requests>=2.31.0
pywin32>=306
